+++ kibot/out_ps_sch_print.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
-# Copyright (c) 2020-2023 Salvador E. Tropea
-# Copyright (c) 2020-2023 Instituto Nacional de Tecnología Industrial
+# Copyright (c) 2020-2026 Salvador E. Tropea
+# Copyright (c) 2020-2026 Instituto Nacional de Tecnología Industrial
 # License: AGPL-3.0
 # Project: KiBot (formerly KiPlot)
 """
@@ -13,7 +13,7 @@ Dependencies:
 """
 from .gs import GS
 from .out_any_sch_print import Any_SCH_PrintOptions
-from .misc import PS_SCH_PRINT, FONT_HELP_TEXT
+from .misc import PS_SCH_PRINT, FONT_HELP_TEXT, MULTI_SCH_NAME
 from .macros import macros, document, output_class  # noqa: F401
 from . import log
 
@@ -36,7 +36,7 @@ class PS_SCH_Print(BaseOutput):  # noqa: F821
         Exports the schematic in postscript. Suitable for printing.
         This output is what you get from the 'File/Plot' menu in eeschema.
         Supports the image replacement using the prefix indicated by the `sch_image_prefix` global variable """
-    __doc__ += FONT_HELP_TEXT
+    __doc__ += MULTI_SCH_NAME+FONT_HELP_TEXT
 
     def __init__(self):
         super().__init__()